+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - set default min python version to 3.7
 - add authors/maintainers in matadata
+- add classifiers in metadata
 
 ## [v0.16.0] - 2022-09-22