+++ CHANGELOG.md
@@ -184,4 +184,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v0.4.0]: https://github.com/cmake-wheel/cmeel/compare/v0.3.0...v0.4.0
 [v0.3.0]: https://github.com/cmake-wheel/cmeel/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/cmake-wheel/cmeel/compare/v0.1.0...v0.2.0
-[v2.0.0]: https://github.com/cmake-wheel/cmeel/releases/tag/v0.1.0
+[v0.1.0]: https://github.com/cmake-wheel/cmeel/releases/tag/v0.1.0